refactor: extract AbstractBatchableGraphics base class
refactor: extract AbstractGraphicsContextRenderData base class

refactor: extract GL/GPU adaptor base classes

refactor: quick fixes and fixOrientation reuse

refactor: extract AbstractGraphicsPipe base class
