+++ src/scene/graphics-smooth/shared/BatchableSmoothGraphics.ts
@@ -1,13 +1,7 @@
-import { Matrix } from '../../../maths/matrix/Matrix';
-import { multiplyHexColors } from '../../container/utils/multiplyHexColors';
+import { AbstractBatchableGraphics } from '../../graphics/shared/AbstractBatchableGraphics';
 
-import type { Batch, BatchableMeshElement, Batcher } from '../../../rendering/batcher/shared/Batcher';
-import type { Topology } from '../../../rendering/renderers/shared/geometry/const';
-import type { BLEND_MODES } from '../../../rendering/renderers/shared/state/const';
-import type { Texture } from '../../../rendering/renderers/shared/texture/Texture';
-import type { SmoothGraphics } from './SmoothGraphics';
-
-const identityMatrix = new Matrix();
+import type { BatchableMeshElement } from '../../../rendering/batcher/shared/Batcher';
+import type { Graphics } from '../../graphics/shared/Graphics';
 
 /**
  * Geometry data for smooth graphics: 10-float-per-vertex interleaved buffer + indices.
@@ -25,24 +19,9 @@ export interface SmoothGeometryData
  * @category scene
  * @advanced
  */
-export class BatchableSmoothGraphics implements BatchableMeshElement
+export class BatchableSmoothGraphics extends AbstractBatchableGraphics<Graphics> implements BatchableMeshElement
 {
-    public readonly packAsQuad = false;
     public batcherName = 'smooth';
-    public topology: Topology = 'triangle-list';
-
-    public texture: Texture;
-    public renderable: SmoothGraphics;
-
-    public indexOffset: number;
-    public indexSize: number;
-    public attributeOffset: number;
-    public attributeSize: number;
-
-    public baseColor: number;
-    public alpha: number;
-    public applyTransform = true;
-    public roundPixels: 0 | 1 = 0;
 
     /** Line width in local space (half-width is computed in packer) */
     public lineWidth = 0;
@@ -53,12 +32,6 @@ export class BatchableSmoothGraphics implements BatchableMeshElement
 
     /** 10-float-per-vertex geometry data */
     public geometryData: SmoothGeometryData;
-
-    public _indexStart: number;
-    public _textureId: number;
-    public _attributeStart: number;
-    public _batcher: Batcher = null;
-    public _batch: Batch = null;
 
     get uvs(): Float32Array
     {
@@ -73,36 +46,6 @@ export class BatchableSmoothGraphics implements BatchableMeshElement
     get indices(): Uint16Array | Uint32Array
     {
         return this.geometryData.indices;
-    }
-
-    get blendMode(): BLEND_MODES
-    {
-        if (this.renderable && this.applyTransform)
-        {
-            return this.renderable.groupBlendMode;
-        }
-
-        return 'normal';
-    }
-
-    get color(): number
-    {
-        const rgb = this.baseColor;
-        const bgr = (rgb >> 16) | (rgb & 0xff00) | ((rgb & 0xff) << 16);
-        const renderable = this.renderable;
-
-        if (renderable)
-        {
-            return multiplyHexColors(bgr, renderable.groupColor)
-                + ((this.alpha * renderable.groupAlpha * 255) << 24);
-        }
-
-        return bgr + ((this.alpha * 255) << 24);
-    }
-
-    get transform(): Matrix
-    {
-        return this.renderable?.groupTransform || identityMatrix;
     }
 
     public copyTo(target: BatchableSmoothGraphics): void
@@ -121,19 +64,9 @@ export class BatchableSmoothGraphics implements BatchableMeshElement
         target.pixelLine = this.pixelLine;
     }
 
-    public reset(): void
+    public override destroy(): void
     {
-        this.applyTransform = true;
-        this.renderable = null;
-        this.topology = 'triangle-list';
-    }
-
-    public destroy(): void
-    {
-        this.renderable = null;
-        this.texture = null;
+        super.destroy();
         this.geometryData = null;
-        this._batcher = null;
-        this._batch = null;
     }
 }
