+++ src/scene/index.ts
@@ -58,11 +58,16 @@ export * from './graphics-smooth/shared/SmoothSegmentPacker';
 export * from './graphics/canvas/CanvasGraphicsAdaptor';
 export * from './graphics/canvas/CanvasGraphicsContextSystem';
 export * from './graphics/canvas/CanvasGraphicsPipe';
+export * from './graphics/gl/AbstractGlGraphicsAdaptor';
 export * from './graphics/gl/GlGraphicsAdaptor';
+export * from './graphics/gpu/AbstractGpuGraphicsAdaptor';
 export * from './graphics/gpu/colorToUniform';
 export * from './graphics/gpu/GpuGraphicsAdaptor';
+export * from './graphics/shared/AbstractBatchableGraphics';
 export * from './graphics/shared/AbstractGpuGraphicsContext';
+export * from './graphics/shared/AbstractGraphicsContextRenderData';
 export * from './graphics/shared/AbstractGraphicsContextSystem';
+export * from './graphics/shared/AbstractGraphicsPipe';
 export * from './graphics/shared/BatchableGraphics';
 export * from './graphics/shared/buildCommands/buildAdaptiveBezier';
 export * from './graphics/shared/buildCommands/buildAdaptiveQuadratic';
